Add extras.php with customized "Read more" link
Switching from our previous theme to this one changed the RSS feed excerpts, which we display on dashboard.wikiedu.org. This change makes the feed excerpts match the format of our previous theme, with a "Continued" link at the end of the excerpt rather than "Read More", and without putting that link into a new paragraph.

This extras.php file is a copy one from the latest version of the understrap base theme, except that the `understrap_all_excerpts_get_more_link` function at the end has been modified.